requirements.txt
discord.py==1.7.2
pynput==1.7.3